feat: Deploy template with external variables (#17)
## Problem

Since PR #16, it's possible to render a template relying on variables (e.g. variant data) loaded from another template.

This works for the `render` CLI command, but not for the `deploy` command. I.e. if you deploy a template which relies on external variables, these variables will not be loaded and the published page will crash because of that.

## Solution

Add support for `LOAD_VARS_FROM_TEMPLATE` env var in the `deploy` CLI command.